Refactor validation interfaces and handlers: consolidate InputHandler and InputValidator, update package structure, and remove unused classes.
