--- src/main/java/utilities/Validator.java
@@ -1,7 +1,0 @@
-package utilities;
-
-public interface Validator<T> {
-    T parse(String input);
-
-    boolean isValid(T value);
-}
