+++ src/main/java/validation/InputHandler.java
@@ -1,4 +1,4 @@
-package utilities;
+package validation;
 
 public interface InputHandler<T> {
     T getInput(String prompt);
