+++ src/main/java/validation/InputValidatorImpl.java
@@ -1,26 +1,37 @@
 package validation;
 
-import utilities.InputValidator;
-
 public class InputValidatorImpl<T> implements InputValidator<T> {
     private final Validator<T> validator;
+    private final String typeName;
 
-    public InputValidatorImpl(Validator<T> validator) {
+    private final String errorMessage;
+
+    public InputValidatorImpl(final Validator<T> validator, final String typeName, final String errorMessage) {
         this.validator = validator;
+        this.typeName = typeName;
+        this.errorMessage = errorMessage;
+    }
+
+    @Override
+    public T parse(final String input) {
+        return this.validator.parse(input);
+    }
+
+    @Override
+    public boolean isValid(final T value) {
+        return this.validator.isValid(value);
     }
 
     @Override
     public String getTypeName() {
-        return this.validator.getTypeName();
+        return this.typeName;
     }
 
     @Override
-    public boolean isValid(T input) {
-        return this.validator.validate(input);
+
+    public String getErrorMessage() {
+        return this.errorMessage;
+
     }
 
-    @Override
-    public T parse(String input) {
-        return this.validator.parse(input);
-    }
 }
