+++ src/main/java/validation/Validator.java
@@ -1,7 +1,7 @@
 package validation;
 
 public interface Validator<T> {
-    boolean validate(T input);
     T parse(String input);
-    String getTypeName();
+
+    boolean isValid(T value);
 }
